+++ scheduler/scheduler.go
@@ -9,12 +9,9 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/apache/arrow/go/v16/arrow"
-
 	"github.com/cloudquery/plugin-sdk/v4/caser"
 	"github.com/cloudquery/plugin-sdk/v4/message"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
-	"github.com/getsentry/sentry-go"
 	"github.com/rs/zerolog"
 	"github.com/thoas/go-funk"
 	"go.opentelemetry.io/otel"
@@ -107,6 +104,9 @@ type Scheduler struct {
 
 	// The maximum number of go routines that can be spawned for a specific resource
 	singleResourceMaxConcurrency int64
+
+	// Controls how records are constructed on the source side.
+	batchSettings *BatchSettings
 }
 
 type syncClient struct {
@@ -126,6 +126,7 @@ func NewScheduler(opts ...Option) *Scheduler {
 		maxDepth:                        DefaultMaxDepth,
 		singleResourceMaxConcurrency:    DefaultSingleResourceMaxConcurrency,
 		singleNestedTableMaxConcurrency: DefaultSingleNestedTableMaxConcurrency,
+		batchSettings:                   new(BatchSettings),
 	}
 	for _, opt := range opts {
 		opt(&s)
@@ -187,7 +188,7 @@ func (s *Scheduler) Sync(ctx context.Context, client schema.ClientMeta, tables s
 	// send migrate messages first
 	for _, tableOriginal := range tables.FlattenTables() {
 		migrateMessage := &message.SyncMigrateTable{
-			Table: tableOriginal.Copy(nil),
+			Table: tableOriginal.Copy(tableOriginal.Parent),
 		}
 		if syncClient.deterministicCQID {
 			schema.CqIDAsPK(migrateMessage.Table)
@@ -209,33 +210,37 @@ func (s *Scheduler) Sync(ctx context.Context, client schema.ClientMeta, tables s
 			panic(fmt.Errorf("unknown scheduler %s", s.strategy.String()))
 		}
 	}()
+
+	b := s.batchSettings.getBatcher(ctx, res, s.logger)
+	defer b.close()    // wait for all resources to be processed
+	done := ctx.Done() // no need to do the lookups in loop
 	for resource := range resources {
 		select {
-		case res <- &message.SyncInsert{Record: resourceToRecord(resource)}:
-		case <-ctx.Done():
+		case <-done:
 			s.logger.Debug().Msg("sync context cancelled")
 			return context.Cause(ctx)
+		default:
+			b.process(resource)
 		}
 	}
 	return context.Cause(ctx)
-}
-
-func resourceToRecord(resource *schema.Resource) arrow.Record {
-	vector := resource.GetValues()
-	return vector.ToArrowRecord(resource.Table.ToArrowSchema())
 }
 
 func (s *syncClient) logTablesMetrics(tables schema.Tables, client Client) {
 	clientName := client.ID()
 	for _, table := range tables {
 		metrics := s.metrics.TableClient[table.Name][clientName]
-		s.logger.Info().Str("table", table.Name).Str("client", clientName).Uint64("resources", metrics.Resources).Uint64("errors", metrics.Errors).Msg("table sync finished")
+		duration := metrics.Duration.Load()
+		if duration == nil {
+			// This can happen for a relation when there are no resources to resolve from the parent
+			duration = new(time.Duration)
+		}
+		s.logger.Info().Str("table", table.Name).Str("client", clientName).Uint64("resources", metrics.Resources).Dur("duration_ms", *duration).Uint64("errors", metrics.Errors).Msg("table sync finished")
 		s.logTablesMetrics(table.Relations, client)
 	}
 }
 
 func (s *syncClient) resolveResource(ctx context.Context, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, item any) *schema.Resource {
-	var validationErr *schema.ValidationError
 	ctx, cancel := context.WithTimeout(ctx, 10*time.Minute)
 	defer cancel()
 	resource := schema.NewResourceData(table, parent, item)
@@ -248,22 +253,12 @@ func (s *syncClient) resolveResource(ctx context.Context, table *schema.Table, c
 			stack := fmt.Sprintf("%s\n%s", err, string(debug.Stack()))
 			logger.Error().Interface("error", err).TimeDiff("duration", time.Now(), objectStartTime).Str("stack", stack).Msg("resource resolver finished with panic")
 			atomic.AddUint64(&tableMetrics.Panics, 1)
-			sentry.WithScope(func(scope *sentry.Scope) {
-				scope.SetTag("table", table.Name)
-				sentry.CurrentHub().CaptureMessage(stack)
-			})
 		}
 	}()
 	if table.PreResourceResolver != nil {
 		if err := table.PreResourceResolver(ctx, client, resource); err != nil {
 			logger.Error().Err(err).Msg("pre resource resolver failed")
 			atomic.AddUint64(&tableMetrics.Errors, 1)
-			if errors.As(err, &validationErr) {
-				sentry.WithScope(func(scope *sentry.Scope) {
-					scope.SetTag("table", table.Name)
-					sentry.CurrentHub().CaptureMessage(validationErr.MaskedError())
-				})
-			}
 			return nil
 		}
 	}
@@ -276,12 +271,6 @@ func (s *syncClient) resolveResource(ctx context.Context, table *schema.Table, c
 		if err := table.PostResourceResolver(ctx, client, resource); err != nil {
 			logger.Error().Stack().Err(err).Msg("post resource resolver finished with error")
 			atomic.AddUint64(&tableMetrics.Errors, 1)
-			if errors.As(err, &validationErr) {
-				sentry.WithScope(func(scope *sentry.Scope) {
-					scope.SetTag("table", table.Name)
-					sentry.CurrentHub().CaptureMessage(validationErr.MaskedError())
-				})
-			}
 		}
 	}
 	atomic.AddUint64(&tableMetrics.Resources, 1)
@@ -289,18 +278,12 @@ func (s *syncClient) resolveResource(ctx context.Context, table *schema.Table, c
 }
 
 func (s *syncClient) resolveColumn(ctx context.Context, logger zerolog.Logger, tableMetrics *TableClientMetrics, client schema.ClientMeta, resource *schema.Resource, c schema.Column) {
-	var validationErr *schema.ValidationError
 	columnStartTime := time.Now()
 	defer func() {
 		if err := recover(); err != nil {
 			stack := fmt.Sprintf("%s\n%s", err, string(debug.Stack()))
 			logger.Error().Str("column", c.Name).Interface("error", err).TimeDiff("duration", time.Now(), columnStartTime).Str("stack", stack).Msg("column resolver finished with panic")
 			atomic.AddUint64(&tableMetrics.Panics, 1)
-			sentry.WithScope(func(scope *sentry.Scope) {
-				scope.SetTag("table", resource.Table.Name)
-				scope.SetTag("column", c.Name)
-				sentry.CurrentHub().CaptureMessage(stack)
-			})
 		}
 	}()
 
@@ -308,13 +291,6 @@ func (s *syncClient) resolveColumn(ctx context.Context, logger zerolog.Logger, t
 		if err := c.Resolver(ctx, client, resource, c); err != nil {
 			logger.Error().Err(err).Msg("column resolver finished with error")
 			atomic.AddUint64(&tableMetrics.Errors, 1)
-			if errors.As(err, &validationErr) {
-				sentry.WithScope(func(scope *sentry.Scope) {
-					scope.SetTag("table", resource.Table.Name)
-					scope.SetTag("column", c.Name)
-					sentry.CurrentHub().CaptureMessage(validationErr.MaskedError())
-				})
-			}
 		}
 	} else {
 		// base use case: try to get column with CamelCase name
@@ -324,13 +300,6 @@ func (s *syncClient) resolveColumn(ctx context.Context, logger zerolog.Logger, t
 			if err != nil {
 				logger.Error().Err(err).Msg("column resolver finished with error")
 				atomic.AddUint64(&tableMetrics.Errors, 1)
-				if errors.As(err, &validationErr) {
-					sentry.WithScope(func(scope *sentry.Scope) {
-						scope.SetTag("table", resource.Table.Name)
-						scope.SetTag("column", c.Name)
-						sentry.CurrentHub().CaptureMessage(validationErr.MaskedError())
-					})
-				}
 			}
 		}
 	}
@@ -348,13 +317,4 @@ func maxDepth(tables schema.Tables) uint64 {
 		}
 	}
 	return depth
-}
-
-// unparam's suggestion to remove the second parameter is not good advice here.
-// nolint:unparam
-func max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
 }